requirements.txt
feedparser
datetime
python-dotenv